requirements.txt
 torch
 torchvision
 dvc
 matplotlib
 numpy
 gitpython
 opencv-python>=4.1.1
 Pillow>=7.1.2
 psutil
 PyYAML>=5.3.1
 requests>=2.23.0
 scipy>=1.4.1
 tqdm>4.64.0
 tensorboard>=2.4.1
 pandas>=1.1.4
 seaborn>=0.11.0
 scikit-learn